# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /api/v1/mysql:
    post:
      parameters:
      - description: Namespace
        in: query
        name: namespace
        required: true
        type: string
      - description: ClusterName
        in: query
        name: cluster_name
        required: true
        type: string
      - description: user
        in: query
        name: user
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      - description: storage_type
        in: query
        name: storage_type
        required: true
        type: string
      - description: multi_master
        in: query
        name: multi_master
        required: true
        type: boolean
      - description: version
        in: query
        name: version
        type: string
      - description: port
        in: query
        name: port
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: add mysql cluster
  /api/v1/mysql/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Delete mysql cluster
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get a single mysql cluster
  /api/v1/mysqls:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get all mysql cluster
swagger: "2.0"